refactor and tidy host and child components
