+++ src/app/shared/components/template/components/base.ts
@@ -5,7 +5,17 @@ import { TemplateContainerComponent } from "../template-container.component";
 @Component({
   selector: "plh-template-component-base",
   template: ``,
+  styleUrls: ["./tmpl-components-common.scss"],
 })
+/**
+ * Common methods and data made available to all other components
+ * By default it provides input setters, and mapping to local variables
+ *
+ * Other components can either extend this one, or implement ITemplateRowProps
+ * in their own way.
+ * Note, if extending the base component access to data is provided by the declared properties,
+ * e.g. `_row`, `_localVariables`
+ */
 export class TemplateBaseComponent implements ITemplateRowProps {
   _row: FlowTypes.TemplateRow;
   _localVariables: { [name: string]: any };
@@ -17,7 +27,12 @@ export class TemplateBaseComponent implements ITemplateRowProps {
   @Input() set localVariables(localVariables: { [name: string]: any }) {
     this._localVariables = localVariables;
   }
-  /** reference to parent template container */
-  @Input() parent: { name: string; component: TemplateContainerComponent };
+  /** reference to parent template container - does not have setter as should remain static */
+  @Input() parent: TemplateContainerComponent;
   constructor() {}
+
+  /** Whenever actions are triggered handle in the parent template component */
+  triggerActions() {
+    this.parent.handleActions(this._row.action_list);
+  }
 }
