+++ src/app/shared/components/template/components/text.ts
@@ -1,19 +1,10 @@
-import { Component, Input, OnInit } from "@angular/core";
-import { FlowTypes } from 'src/app/shared/model/flowTypes';
-import { ITemplateComponent } from "./tmpl.component";
+import { Component, OnInit } from "@angular/core";
+import { BaseComponent } from "@fullcalendar/angular";
+import { TemplateBaseComponent } from "./base";
 
 @Component({
   selector: "plh-tmpl-text",
-  template: `<p>{{row.value | localVarsReplace: localVariables}}</p>`,
-  styleUrls: ["./tmpl-components-common.scss"]
+  template: `<p>{{ _row.value | localVarsReplace: localVariables }}</p>`,
+  styleUrls: ["./tmpl-components-common.scss"],
 })
-export class TmplTextComponent implements ITemplateComponent, OnInit {
-
-  @Input() row: FlowTypes.TemplateRow;
-  @Input() template: FlowTypes.Template;
-  @Input() localVariables: { [name: string]: string};
-
-  constructor() {}
-
-  ngOnInit() {}
-}
+export class TmplTextComponent extends TemplateBaseComponent {}
