+++ src/app/shared/components/template/models/index.ts
@@ -2,13 +2,6 @@ import { FlowTypes } from "scripts/types";
 import { TemplateContainerComponent } from "../template-container.component";
 export { FlowTypes } from "scripts/types";
 
-// export interface ITemplateComponent {
-//   row?: FlowTypes.TemplateRow;
-//   rows?: FlowTypes.TemplateRow[];
-//   localVariables?: { [name: string]: any };
-//   parent?: { name: string; component: ITemplateComponent };
-// }
-
 /**
  * Properties passed to a template container instance
  * @param name - flow_name of template to lookup
@@ -19,11 +12,14 @@ export interface ITemplateContainerProps {
   parent?: { name: string; component?: any };
 }
 
+/**
+ * Properties passed to a template row instance
+ * @param row specific data used in component rendering
+ * @param localVariables compiled list of variables used across all template rows
+ * @param parent reference to parent template container
+ */
 export interface ITemplateRowProps {
-  /** specific data used in component rendering */
   row: FlowTypes.TemplateRow;
-  /** compiled list of variables used across all template rows */
   localVariables: { [name: string]: any };
-  /** reference to parent template container */
-  parent: { name: string; component: TemplateContainerComponent };
+  parent: TemplateContainerComponent;
 }
