+++ src/app/shared/components/template/template-container.component.html
@@ -14,6 +14,6 @@
     class="row"
     [row]="row"
     [localVariables]="localVariables"
-    [parent]="parent"
+    [parent]="this"
   ></plh-template-component>
 </div>
